move hard coded metrics to its own package
